backend/requirements.txt
environs==14.2.0
fastapi==0.116.1
fastapi_login==1.10.3
icecream==2.1.3
passlib==1.7.4
pydantic==2.11.7
tortoise-orm[asyncpg]
uvicorn
aerich
pydantic[email]
python-multipart
argon2_cffi
